restructure files and add input validation
